Miscelaneous fixes (#271)
* url: expand size to support longer urls

Services like Brewersfriend use the token as part of the url, which
greatly increases the length of the url to post to. Increase the field
to 80 (TKIDSIZE * 2) to support such services. As a reference,
Brewerfriends url seems to be at least 51 characters long.

* docs/build: fix build instructions

`pio run` should be called from the top-level directory.